requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
cryptocompare==0.7.6
dateparser==1.2.0
frozendict==2.4.4
frozenlist==1.4.1
html5lib==1.1
idna==3.7
influxdb-client==1.44.0
lxml==5.2.2
multidict==6.0.5
multitasking==0.0.11
numpy==2.0.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pyaml==24.7.0
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
reactivex==4.0.4
regex==2024.7.24
requests==2.32.3
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
yfinance==0.2.41
